Move Vertices of Rectangle and Circle with MASSIVE BUG
If a vertex is moved beyond the initial boundaries of the bounding rectangle, the vertex is no longer e.g. the lower right one but turns in the upper right one. This alters the order within the points drastically and as of yet i dont know how to fix it